Moved MicrolabInterface to new file
